+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html
@@ -93,7 +93,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">org.opengis.cite.cdb10.CDBStructure</div>
+<div class="subTitle">org.opengis.cite.cdb10.cdbStructure</div>
 <h2 title="Class Capability1Tests" class="title">Class Capability1Tests</h2>
 </div>
 <div class="contentContainer">
@@ -104,7 +104,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">org.opengis.cite.cdb10.CommonFixture</a></li>
 <li>
 <ul class="inheritance">
-<li>org.opengis.cite.cdb10.CDBStructure.Capability1Tests</li>
+<li>org.opengis.cite.cdb10.cdbStructure.Capability1Tests</li>
 </ul>
 </li>
 </ul>
