+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/CDBStructure/class-use/Capability1Tests.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class org.opengis.cite.cdb10.CDBStructure.Capability1Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
+<title>Uses of Class org.opengis.cite.cdb10.cdbStructure.Capability1Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../../script.js"></script>
@@ -13,13 +13,12 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class org.opengis.cite.cdb10.CDBStructure.Capability1Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
+            parent.document.title="Uses of Class org.opengis.cite.cdb10.cdbStructure.Capability1Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
         }
     }
     catch(err) {
     }
-//-->
-</script>
+//--></script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
@@ -34,7 +33,7 @@
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../../overview-summary.html">Overview</a></li>
 <li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Class</a></li>
+<li><a href="../../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.cdbStructure">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../package-tree.html">Tree</a></li>
 <li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
@@ -71,9 +70,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class org.opengis.cite.cdb10.CDBStructure.Capability1Tests" class="title">Uses of Class<br>org.opengis.cite.cdb10.CDBStructure.Capability1Tests</h2>
+<h2 title="Uses of Class org.opengis.cite.cdb10.cdbStructure.Capability1Tests" class="title">Uses of Class<br>org.opengis.cite.cdb10.cdbStructure.Capability1Tests</h2>
 </div>
-<div class="classUseContainer">No usage of org.opengis.cite.cdb10.CDBStructure.Capability1Tests</div>
+<div class="classUseContainer">No usage of org.opengis.cite.cdb10.cdbStructure.Capability1Tests</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
@@ -85,7 +84,7 @@
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../../overview-summary.html">Overview</a></li>
 <li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Class</a></li>
+<li><a href="../../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.cdbStructure">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../package-tree.html">Tree</a></li>
 <li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
