+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/CDBStructure/package-tree.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>org.opengis.cite.cdb10.CDBStructure Class Hierarchy (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
+<title>org.opengis.cite.cdb10.cdbStructure Class Hierarchy (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -13,13 +13,12 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="org.opengis.cite.cdb10.CDBStructure Class Hierarchy (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
+            parent.document.title="org.opengis.cite.cdb10.cdbStructure Class Hierarchy (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
         }
     }
     catch(err) {
     }
-//-->
-</script>
+//--></script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
@@ -71,7 +70,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 class="title">Hierarchy For Package org.opengis.cite.cdb10.CDBStructure</h1>
+<h1 class="title">Hierarchy For Package org.opengis.cite.cdb10.cdbStructure</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../../../../overview-tree.html">All Packages</a></li>
@@ -84,7 +83,7 @@
 <ul>
 <li type="circle">org.opengis.cite.cdb10.<a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10"><span class="typeNameLink">CommonFixture</span></a>
 <ul>
-<li type="circle">org.opengis.cite.cdb10.CDBStructure.<a href="../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure"><span class="typeNameLink">Capability1Tests</span></a></li>
+<li type="circle">org.opengis.cite.cdb10.cdbStructure.<a href="../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.cdbStructure"><span class="typeNameLink">Capability1Tests</span></a></li>
 </ul>
 </li>
 </ul>
