+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/class-use/CommonFixture.html
@@ -84,13 +84,13 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.CDBStructure">org.opengis.cite.cdb10.CDBStructure</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.cdbStructure">org.opengis.cite.cdb10.cdbStructure</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 1 includes basic facilities.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.MetadataAndVersioning">org.opengis.cite.cdb10.MetadataAndVersioning</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.metadataAndVersioning">org.opengis.cite.cdb10.metadataAndVersioning</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </td>
@@ -100,12 +100,12 @@
 </li>
 <li class="blockList">
 <ul class="blockList">
-<li class="blockList"><a name="org.opengis.cite.cdb10.CDBStructure">
+<li class="blockList"><a name="org.opengis.cite.cdb10.cdbStructure">
 <!--   -->
 </a>
-<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.CDBStructure</a></h3>
+<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.cdbStructure</a></h3>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing subclasses, and an explanation">
-<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.CDBStructure</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.cdbStructure</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colFirst" scope="col">Modifier and Type</th>
 <th class="colLast" scope="col">Class and Description</th>
@@ -113,19 +113,19 @@
 <tbody>
 <tr class="altColor">
 <td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Capability1Tests</a></span></code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.cdbStructure">Capability1Tests</a></span></code>
 <div class="block">Includes various tests of capability 1.</div>
 </td>
 </tr>
 </tbody>
 </table>
 </li>
-<li class="blockList"><a name="org.opengis.cite.cdb10.MetadataAndVersioning">
+<li class="blockList"><a name="org.opengis.cite.cdb10.metadataAndVersioning">
 <!--   -->
 </a>
-<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></h3>
+<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.metadataAndVersioning</a></h3>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing subclasses, and an explanation">
-<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.metadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colFirst" scope="col">Modifier and Type</th>
 <th class="colLast" scope="col">Class and Description</th>
@@ -133,7 +133,7 @@
 <tbody>
 <tr class="altColor">
 <td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></span></code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning">Capability2Tests</a></span></code>
 <div class="block">Includes various tests of capability 2.</div>
 </td>
 </tr>
