+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/level2/class-use/Capability2Tests.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests (Test suite: ets-cdb10 0.1-SNAPSHOT
+<title>Uses of Class org.opengis.cite.cdb10.metadataAndVersioning.Capability2Tests (Test suite: ets-cdb10 0.1-SNAPSHOT
     API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
@@ -14,7 +14,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
+            parent.document.title="Uses of Class org.opengis.cite.cdb10.metadataAndVersioning.Capability2Tests (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
         }
     }
     catch(err) {
@@ -34,7 +34,7 @@
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../../overview-summary.html">Overview</a></li>
 <li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Class</a></li>
+<li><a href="../../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../package-tree.html">Tree</a></li>
 <li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
@@ -71,9 +71,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests" class="title">Uses of Class<br>org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests</h2>
+<h2 title="Uses of Class org.opengis.cite.cdb10.metadataAndVersioning.Capability2Tests" class="title">Uses of Class<br>org.opengis.cite.cdb10.metadataAndVersioning.Capability2Tests</h2>
 </div>
-<div class="classUseContainer">No usage of org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests</div>
+<div class="classUseContainer">No usage of org.opengis.cite.cdb10.metadataAndVersioning.Capability2Tests</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
@@ -85,7 +85,7 @@
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../../overview-summary.html">Overview</a></li>
 <li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Class</a></li>
+<li><a href="../../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../package-tree.html">Tree</a></li>
 <li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
