+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/level2/package-frame.html
@@ -4,17 +4,17 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>org.opengis.cite.cdb10.MetadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
+<title>org.opengis.cite.cdb10.metadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html" target="classFrame">org.opengis.cite.cdb10.MetadataAndVersioning</a></h1>
+<h1 class="bar"><a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html" target="classFrame">org.opengis.cite.cdb10.metadataAndVersioning</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning" target="classFrame">Capability2Tests</a></li>
+<li><a href="Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning" target="classFrame">Capability2Tests</a></li>
 </ul>
 </div>
 </body>
