+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/level2/package-summary.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>org.opengis.cite.cdb10.MetadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
+<title>org.opengis.cite.cdb10.metadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="org.opengis.cite.cdb10.MetadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
+            parent.document.title="org.opengis.cite.cdb10.metadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
         }
     }
     catch(err) {
@@ -70,7 +70,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;org.opengis.cite.cdb10.MetadataAndVersioning</h1>
+<h1 title="Package" class="title">Package&nbsp;org.opengis.cite.cdb10.metadataAndVersioning</h1>
 <div class="docSummary">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </div>
@@ -87,7 +87,7 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></td>
+<td class="colFirst"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning">Capability2Tests</a></td>
 <td class="colLast">
 <div class="block">Includes various tests of capability 2.</div>
 </td>
@@ -99,8 +99,8 @@
 <a name="package.description">
 <!--   -->
 </a>
-<h2 title="Package org.opengis.cite.cdb10.MetadataAndVersioning Description">Package
-    org.opengis.cite.cdb10.MetadataAndVersioning Description</h2>
+<h2 title="Package org.opengis.cite.cdb10.metadataAndVersioning Description">Package
+    org.opengis.cite.cdb10.metadataAndVersioning Description</h2>
 <div class="block">Conformance Level 2 extends Level 1. The following additional capabilities
  are required: 
  <ul>
