+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/level2/package-use.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Wed May 25 00:42:42 EDT 2016 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Package org.opengis.cite.cdb10.MetadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
+<title>Uses of Package org.opengis.cite.cdb10.metadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)</title>
 <meta name="date" content="2016-05-25">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Package org.opengis.cite.cdb10.MetadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
+            parent.document.title="Uses of Package org.opengis.cite.cdb10.metadataAndVersioning (Test suite: ets-cdb10 0.1-SNAPSHOT API)";
         }
     }
     catch(err) {
@@ -70,9 +70,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Uses of Package org.opengis.cite.cdb10.MetadataAndVersioning" class="title">Uses of Package<br>org.opengis.cite.cdb10.MetadataAndVersioning</h1>
+<h1 title="Uses of Package org.opengis.cite.cdb10.metadataAndVersioning" class="title">Uses of Package<br>org.opengis.cite.cdb10.metadataAndVersioning</h1>
 </div>
-<div class="contentContainer">No usage of org.opengis.cite.cdb10.MetadataAndVersioning</div>
+<div class="contentContainer">No usage of org.opengis.cite.cdb10.metadataAndVersioning</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
