+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/package-use.html
@@ -91,13 +91,13 @@
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.CDBStructure">org.opengis.cite.cdb10.CDBStructure</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.cdbStructure">org.opengis.cite.cdb10.cdbStructure</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 1 includes basic facilities.</div>
 </td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.MetadataAndVersioning">org.opengis.cite.cdb10.MetadataAndVersioning</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.metadataAndVersioning">org.opengis.cite.cdb10.metadataAndVersioning</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </td>
@@ -128,34 +128,34 @@
 </tbody>
 </table>
 </li>
-<li class="blockList"><a name="org.opengis.cite.cdb10.CDBStructure">
+<li class="blockList"><a name="org.opengis.cite.cdb10.cdbStructure">
 <!--   -->
 </a>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing classes, and an explanation">
-<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.CDBStructure</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.cdbStructure</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colOne" scope="col">Class and Description</th>
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.CDBStructure">CommonFixture</a>
+<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.cdbStructure">CommonFixture</a>
 <div class="block">A supporting base class that sets up a common test fixture.</div>
 </td>
 </tr>
 </tbody>
 </table>
 </li>
-<li class="blockList"><a name="org.opengis.cite.cdb10.MetadataAndVersioning">
+<li class="blockList"><a name="org.opengis.cite.cdb10.metadataAndVersioning">
 <!--   -->
 </a>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing classes, and an explanation">
-<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.metadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colOne" scope="col">Class and Description</th>
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.MetadataAndVersioning">CommonFixture</a>
+<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.metadataAndVersioning">CommonFixture</a>
 <div class="block">A supporting base class that sets up a common test fixture.</div>
 </td>
 </tr>
