+++ cdb10/1.0/site/apidocs/overview-summary.html
@@ -95,7 +95,7 @@
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.CDBStructure</a></td>
+<td class="colFirst"><a href="org/opengis/cite/cdb10/CDBStructure/package-summary.html">org.opengis.cite.cdb10.cdbStructure</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 1 includes basic facilities.</div>
 </td>
