+++ cdb10/1.0/site/apidocs/overview-tree.html
@@ -75,8 +75,8 @@
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="org/opengis/cite/cdb10/package-tree.html">org.opengis.cite.cdb10</a>, </li>
-<li><a href="org/opengis/cite/cdb10/CDBStructure/package-tree.html">org.opengis.cite.cdb10.CDBStructure</a>, </li>
-<li><a href="org/opengis/cite/cdb10/level2/package-tree.html">org.opengis.cite.cdb10.MetadataAndVersioning</a>, </li>
+<li><a href="org/opengis/cite/cdb10/CDBStructure/package-tree.html">org.opengis.cite.cdb10.cdbStructure</a>, </li>
+<li><a href="org/opengis/cite/cdb10/level2/package-tree.html">org.opengis.cite.cdb10.metadataAndVersioning</a>, </li>
 <li><a href="org/opengis/cite/cdb10/util/package-tree.html">org.opengis.cite.cdb10.util</a></li>
 </ul>
 </div>
@@ -93,8 +93,8 @@
 <li type="circle">org.opengis.cite.cdb10.util.<a href="org/opengis/cite/cdb10/util/ClientUtils.html" title="class in org.opengis.cite.cdb10.util"><span class="typeNameLink">ClientUtils</span></a></li>
 <li type="circle">org.opengis.cite.cdb10.<a href="org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10"><span class="typeNameLink">CommonFixture</span></a>
 <ul>
-<li type="circle">org.opengis.cite.cdb10.CDBStructure.<a href="org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure"><span class="typeNameLink">Capability1Tests</span></a></li>
-<li type="circle">org.opengis.cite.cdb10.MetadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning"><span class="typeNameLink">Capability2Tests</span></a></li>
+<li type="circle">org.opengis.cite.cdb10.cdbStructure.<a href="org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.cdbStructure"><span class="typeNameLink">Capability1Tests</span></a></li>
+<li type="circle">org.opengis.cite.cdb10.metadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.metadataAndVersioning"><span class="typeNameLink">Capability2Tests</span></a></li>
 </ul>
 </li>
 <li type="circle">org.opengis.cite.cdb10.<a href="org/opengis/cite/cdb10/ErrorMessage.html" title="class in org.opengis.cite.cdb10"><span class="typeNameLink">ErrorMessage</span></a></li>
