+++ cdb10/1.0/site/apidocs/package-list
@@ -1,4 +1,4 @@
 org.opengis.cite.cdb10
-org.opengis.cite.cdb10.CDBStructure
-org.opengis.cite.cdb10.MetadataAndVersioning
+org.opengis.cite.cdb10.cdbStructure
+org.opengis.cite.cdb10.metadataAndVersioning
 org.opengis.cite.cdb10.util
