+++ src/main/java/org/opengis/cite/cdb10/cdbStructure/package-info.java
@@ -11,4 +11,4 @@
  * href="http://www.w3.org/TR/html5/infrastructure.html#conformance-classes">HTML5
  * - Conformance classes</a>
  */
-package org.opengis.cite.cdb10.CDBStructure;
+package org.opengis.cite.cdb10.cdbStructure;
