+++ src/main/java/org/opengis/cite/cdb10/metadataAndVersioning/package-info.java
@@ -10,4 +10,4 @@
  * href="http://www.w3.org/TR/html5/infrastructure.html#conformance-classes">HTML5
  * - Conformance classes</a>
  */
-package org.opengis.cite.cdb10.MetadataAndVersioning;
+package org.opengis.cite.cdb10.metadataAndVersioning;
